fix(packaged): install stdio guard at packaged entry bootstrap
Module load runs before main(), so the guard covers every pre-logger
diagnostic (launcher inspection, download attribution) rather than only
the ones after the structured logger exists. Move the first-launch
regression test into its own file.
